third_party/blink/web_tests/fast/inline/continuation-inlines-inserted-in-reverse-after-block-expected.txt

You should see two lines, both with 1,2,3.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('3').offsetLeft is >= document.getElementById('2').offsetLeft
PASS document.getElementById('2').offsetLeft is >= document.getElementById('1').offsetLeft
PASS successfullyParsed is true

TEST COMPLETE

1,2,3,
1,2,3,
